analyticsServer/requirements.txt
grpcio-tools
grpcio
googleapis-common-protos
mysql-connector-python